refactor(entities): expose client api via entrypoints
