+++ src/entities/review/index.js
@@ -1,1 +1,2 @@
+export * from './api/client';
 export { avgRating, clampRating } from './model/selectors';
